merge_0515
뭐 많음

+++ Assets/Scripts/MonsterWeaponDamage.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public class MonsterWeaponDamage : MonoBehaviour
 {
-    public int m_damage;
+    public float m_damage;
 }
